% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{$.metaflow.flowspec.FlowSpec}
\alias{$.metaflow.flowspec.FlowSpec}
\title{Overload getter for self object}
\usage{
\method{$}{metaflow.flowspec.FlowSpec}(self, name)
}
\arguments{
\item{self}{the metaflow self object for each step function}

\item{name}{attribute name}
}
\description{
Overload getter for self object
}
\section{Usage}{

\preformatted{
 print(self$var)
}
}

