% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current}
\alias{current}
\title{Helper utility to access current IDs of interest}
\usage{
current(value)
}
\arguments{
\item{value}{one of flow_name, run_id, origin_run_id,
step_name, task_id, pathspec, namespace,
username, retry_count}
}
\description{
Helper utility to access current IDs of interest
}
\examples{
\dontrun{
current("flow_name")
}
}
