% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{decorator}
\alias{decorator}
\title{Metaflow Decorator.}
\usage{
decorator(x, ...)
}
\arguments{
\item{x}{Type of decorator (e.g, resources, catch, retry, timeout, batch ...)}

\item{...}{Named arguments for the decorator (e.g, cpu=1, memory=1000 ...). Note that memory unit is in MB.}
}
\value{
A decorator object
}
\description{
Decorates the \code{step} with the parameters present in its arguments.
}
\details{
For this method to work properly, the arguments \code{...} should be named, and decorator
type should be the first argument.
}
\section{Usage}{

\preformatted{
decorator("catch", print_exception=FALSE)
decorator("resources", cpu=2, memory=10000)
}
}

