% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_inputs}
\alias{gather_inputs}
\title{Helper utility to gather inputs in a join step}
\usage{
gather_inputs(inputs, input)
}
\arguments{
\item{inputs}{inputs from parent branches}

\item{input}{field to extract from inputs from
parent branches into vector}
}
\description{
Helper utility to gather inputs in a join step
}
\section{usage}{

\preformatted{
gather_inputs(inputs, "alpha")
}
}

