% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns the current Metadata provider.}
\usage{
get_metadata()
}
\value{
String type. Information about the Metadata provider currently selected.
This information typically returns provider specific information (like URL for remote
providers or local paths for local providers.
}
\description{
This call returns the current Metadata being used to return information
about Metaflow objects. If this is not set explicitly using metadata(), the default value is
determined through environment variables.
}
