% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install}
\alias{install}
\title{Install Metaflow python dependencies}
\usage{
install(user_install = TRUE, upgrade = FALSE)
}
\arguments{
\item{user_install}{Whether or not to install into the user directory for pip install. Default to TRUE.}

\item{upgrade}{Whether or not to upgrade metaflow python package. Default to FALSE.}
}
\description{
Install Metaflow python dependencies
}
