% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_metaflow}
\alias{install_metaflow}
\title{Install Metaflow Python package}
\usage{
install_metaflow(
  method = c("conda", "virtualenv"),
  envname = "metaflow",
  version = NULL,
  restart_session = TRUE,
  ...
)
}
\arguments{
\item{method}{`character`, indicates to use `"conda"` or `"virtualenv"`}

\item{envname}{`character`, name of environment into which to install}

\item{version}{`character`, version of Metaflow to install. For general use of this package,
this is set automatically, so you should not need to specify this.}

\item{...}{other arguments sent to [reticulate::conda_install()] or
[reticulate::virtualenv_install()]}
}
\description{
This function wraps installation functions from [reticulate][reticulate::reticulate] to install the Python packages
**metaflow** and it's Python dependencies.
}
\details{
This package uses the [reticulate][reticulate::reticulate] package
to make an interface with the [Metaflow](https://metaflow.org/)
Python package.
}
\examples{
\dontrun{
# not run because it requires Python
install_metaflow()
}
}
\seealso{
[altiar: Installation](https://vegawidget.github.io/metaflow/articles/installation.html),
[reticulate: Using reticulate in an R Package](https://rstudio.github.io/reticulate/articles/package.html),
[reticulate: Installing Python Packages](https://rstudio.github.io/reticulate/articles/python_packages.html)
}
