% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_artifacts}
\alias{merge_artifacts}
\title{Helper utility to merge artifacts in a join step}
\usage{
merge_artifacts(flow, inputs, exclude = list())
}
\arguments{
\item{flow}{flow object}

\item{inputs}{inputs from parent branches}

\item{exclude}{list of artifact names to exclude from merging}
}
\description{
Helper utility to merge artifacts in a join step
}
\examples{
\dontrun{
merge_artifacts(flow, inputs)
}
\dontrun{
merge_artifacts(flow, inputs, list("alpha"))
}
}
