% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaflow-package.R
\docType{package}
\name{metaflow-package}
\alias{metaflow}
\alias{metaflow-package}
\title{metaflow: R Interface to 'Metaflow'}
\description{
Interface to 'Metaflow' \url{https://metaflow.org/}, a human-friendly Python framework for building and managing real-life data science projects. Originally developed at Netflix, Metaflow offers a unified API for data science workflows, covering the full development lifecycle of a data science project. Key features include rapid prototyping, built-in versioning, horizontal and vertical scalability, dependency management, and easy integration with production-grade orchestrators. The R interface allows data scientists using R to leverage Metaflow's scalability and reproducibility while staying within their preferred ecosystem
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metaflow.org/}
  \item \url{https://docs.metaflow.org/v/r/}
  \item \url{https://github.com/Netflix/metaflow}
  \item Report bugs at \url{https://github.com/Netflix/metaflow/issues}
}

}
\author{
\strong{Maintainer}: Bryan Galvin \email{bcgalvin@gmail.com}

}
\keyword{internal}
