% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{metaflow}
\alias{metaflow}
\title{Instantiate a flow}
\usage{
metaflow(cls, ...)
}
\arguments{
\item{cls}{flow class name}

\item{...}{flow decorators}
}
\value{
flow object
}
\description{
Instantiate a flow
}
\section{Usage}{

\preformatted{
metaflow("HelloFlow")
}
}

