% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaflow_location}
\alias{metaflow_location}
\title{Return installation path of metaflow R library}
\usage{
metaflow_location(flowRDS)
}
\arguments{
\item{flowRDS}{path of the RDS file containing the flow object}
}
\description{
Return installation path of metaflow R library
}
