% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mf_deserialize}
\alias{mf_deserialize}
\title{Helper utility to deserialize objects from metaflow
data format to R object}
\usage{
mf_deserialize(object)
}
\arguments{
\item{object}{object to deserialize}
}
\value{
R object
}
\description{
Helper utility to deserialize objects from metaflow
data format to R object
}
