% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_client.R
\name{new_flow}
\alias{new_flow}
\title{Instantiates a new flow object.}
\usage{
new_flow(flow_id)
}
\arguments{
\item{flow_id}{Flow identifier.}
}
\value{
\code{flow} object corresponding to the supplied identifier.
}
\description{
Instantiates a new flow object.
}
