% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_client.R
\name{new_run}
\alias{new_run}
\title{Instantiates a new run object.}
\usage{
new_run(flow_id, run_id)
}
\arguments{
\item{flow_id}{Flow identifier.}

\item{run_id}{Run identifier.}
}
\value{
\code{run} object corresponding to the supplied identifiers.
}
\description{
Instantiates a new run object.
}
