% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_client.R
\name{new_step}
\alias{new_step}
\title{Instantiates a new step object.}
\usage{
new_step(flow_id, run_id, step_id)
}
\arguments{
\item{flow_id}{Flow identifier.}

\item{run_id}{Run identifier.}

\item{step_id}{Step identifier.}
}
\value{
\code{step} object corresponding to the supplied identifiers.
}
\description{
Instantiates a new step object.
}
