% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_client.R
\name{new_task}
\alias{new_task}
\title{Instantiates a new task object.}
\usage{
new_task(flow_id, run_id, step_id, task_id)
}
\arguments{
\item{flow_id}{Flow identifier.}

\item{run_id}{Run identifier.}

\item{step_id}{Step identifier.}

\item{task_id}{Task identifier.}
}
\value{
\code{task} object corresponding to the supplied identifiers.
}
\description{
Instantiates a new task object.
}
