% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_dependencies}
\alias{py_dependencies}
\title{Python packages needed for R. 
numpy is needed to handle native R matrix
pandas is needed to handle R data frame}
\usage{
py_dependencies()
}
\description{
Python packages needed for R. 
numpy is needed to handle native R matrix
pandas is needed to handle R data frame
}
