% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{py_hash}
\alias{py_hash}
\title{Hash an R function}
\usage{
py_hash(fun)
}
\arguments{
\item{fun}{Function}
}
\value{
Positive integer
}
\description{
There is no support for hashing in base R. The \verb{digest} package is
usually used, but in order to avoid introduing another dependency we call the
inbuilt Python hash function. To avoid hyphenating function names, we return
the absolute of the hash. Note also that the hash will change between
sessions (see
\url{https://docs.python.org/3/reference/datamodel.html#object.__hash__}).
}
\details{
The hashes produced by this function won't necessarily match those produced
by Python. Base R supports only 32-bit integers, so anything beyond 32 bits
will overflow.
}
\keyword{internal}
