% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{remove_metaflow}
\alias{remove_metaflow}
\title{Remove Metaflow Python package.}
\usage{
remove_metaflow(method = c("conda", "virtualenv"), prompt = TRUE)
}
\arguments{
\item{method}{`character`, indicates to use `"conda"` or `"virtualenv"`. Default to 'conda'.}

\item{prompt}{`bool`, whether to ask for user prompt before removal. Default to TRUE.}
}
\description{
Remove Metaflow Python package.
}
\examples{
\dontrun{
# not run because it requires Python
remove_metaflow(method='conda')
remove_metaflow(method='virtualenv')
}
}
