% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{remove_metaflow_env}
\alias{remove_metaflow_env}
\title{Remove Metaflow Python package.}
\usage{
remove_metaflow_env(prompt = TRUE)
}
\arguments{
\item{prompt}{\code{bool}, whether to ask for user prompt before removal. Default to TRUE.}
}
\description{
Remove Metaflow Python package.
}
\examples{
\dontrun{
# not run because it requires Python
remove_metaflow_env()
}
}
