% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{reset_default_metadata}
\alias{reset_default_metadata}
\title{Resets the Metadata provider to the default value.}
\usage{
reset_default_metadata()
}
\value{
String type. The result of get_metadata() after resetting the provider.
}
\description{
The default value of the Metadata provider is determined through a
combination of environment variables.
}
