% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{Switch Metadata provider}
\usage{
set_metadata(ms = NULL)
}
\arguments{
\item{ms}{string. Can be a path (selects local metadata), a URL starting with http (selects
the service metadata) or an explicit specification {metadata_type}@{info}; as an
example, you can specify local@{path} or service@{url}.}
}
\value{
a string of the description of the metadata selected
}
\description{
This call has a global effect.
Selecting the local metadata will, for example, not allow access to information
stored in remote metadata providers
}
