% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{set_namespace}
\alias{set_namespace}
\title{Switch to a namespace specified by the given tag.}
\usage{
set_namespace(ns = NULL)
}
\arguments{
\item{ns}{namespace}
}
\description{
Switch to a namespace specified by the given tag.
}
\details{
NULL maps to global namespace.
}
