% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metaflow-package}
\alias{metaflow}
\alias{metaflow-package}
\title{R interface to Metaflow}
\description{
Interface to 'Metaflow' \url{https://metaflow.org/}, a human-friendly
Python framework for building and managing real-life data science projects.
Originally developed at Netflix, Metaflow offers a unified API for data
science workflows, covering the full development lifecycle of a data science project.
}
\details{
The Metaflow package allows data scientists using R to leverage Metaflow's scalability and reproducibility
all while staying within their preferred ecosystem.

See the package website at \url{https://docs.metaflow.org} for complete documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metaflow.org/}
  \item \url{https://docs.metaflow.org/v/r/}
  \item \url{https://github.com/Netflix/metaflow}
  \item Report bugs at \url{https://github.com/Netflix/metaflow/issues}
}

}
\author{
\strong{Maintainer}: Bryan Galvin \email{bcgalvin@gmail.com}

}
