% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metaflow-package}
\alias{metaflow-package}
\alias{_PACKAGE}
\alias{metaflow-r}
\title{metaflow: Metaflow for R-Lang}
\description{
R binding for Metaflow. Metaflow is a human-friendly Python/R library
that helps scientists and engineers build and manage real-life data science projects.
Metaflow was originally developed at Netflix to boost productivity of data scientists
who work on a wide variety of projects from classical statistics to state-of-the-art deep learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metaflow.org/}
  \item \url{https://docs.metaflow.org/}
  \item \url{https://github.com/Netflix/metaflow}
  \item Report bugs at \url{https://github.com/Netflix/metaflow/issues}
}

}
