% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mf_serialize}
\alias{mf_serialize}
\title{Helper utility to serialize R object to metaflow
data format}
\usage{
mf_serialize(object)
}
\arguments{
\item{object}{object to serialize}
}
\value{
metaflow data format object
}
\description{
Helper utility to serialize R object to metaflow
data format
}
